<Refactor> 절대 경로 설정, 폴더 구조 변경
