+++ pet-plate-front/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import StyledComponentsRegistry from './lib/registry'
+import StyledComponentsRegistry from '@lib/registry'
 
  
 export default function RootLayout({
