+++ pet-plate-front/src/lib/registry.tsx
@@ -3,8 +3,8 @@
 import React, { useState } from 'react'
 import { useServerInsertedHTML } from 'next/navigation'
 import { ServerStyleSheet, StyleSheetManager, ThemeProvider } from 'styled-components'
-import GlobalStyles from '../style/GlobalStyles'
-import theme from '../style/theme'
+import GlobalStyles from '@style/GlobalStyles'
+import theme from '@style/theme'
  
 export default function StyledComponentsRegistry({
   children,
