+++ pet-plate-front/src/style/GlobalStyles.tsx
