+++ pet-plate-front/src/style/theme.tsx
