+++ pet-plate-front/tsconfig.json
@@ -18,11 +18,14 @@
         "name": "next"
       }
     ],
+    "baseUrl": "./src",
     "paths": {
-      "@/*": ["./src/*"]
+      "@app/*": ["./app/*"],
+      "@components/*": ["./components/*"],
+      "@lib/*": ["./lib/*"],
+      "@style/*": ["./style/*"]
     },
-    "types": ["styled-components"],
-    "baseUrl": "./src"
+    "types": ["styled-components"]
   },
 
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
